refactor(discord): rename login helper
