+++ discord/lib/handlers/index.js
@@ -2,12 +2,12 @@ import {
   getNewTokenWithDependencies,
   getReuseSessionTokenWithDependencies,
 } from './authenticated';
-import { getRenderLoginWithData } from './login';
+import { getRenderLoginWithData } from './get-render-login-with-data';
 import { handleError } from './handle-error';
 
 export {
-  getRenderLoginWithData,
   getNewTokenWithDependencies,
+  getRenderLoginWithData,
   getReuseSessionTokenWithDependencies,
   handleError,
 };
